+++ app/src/main/AndroidManifest.xml
@@ -18,6 +18,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity android:name=".DaumAddressActivity"></activity>
+        <activity android:name=".PaymentActivity"></activity>
         <activity
             android:name=".ItemUpdateActivity"
             android:windowSoftInputMode="stateAlwaysHidden"></activity>
