+++ app/src/main/java/com/naram/ewoman_project/OrderActivity.java
@@ -270,6 +270,8 @@ public class OrderActivity extends AppCompatActivity {
 
                     }
 
+                    startActivity(intent);
+
                 }
             }
         });
@@ -600,117 +602,6 @@ public class OrderActivity extends AppCompatActivity {
                 errorString = e.toString();
 
                 return null;
-            }
-
-        }
-    }
-
-    class InsertOrderData extends AsyncTask<String, Void, String> {
-
-        ProgressDialog progressDialog;
-
-        @Override
-        protected void onPreExecute() {
-            super.onPreExecute();
-
-            progressDialog = new ProgressDialog(OrderActivity.this, R.style.AlertDialogStyle);
-            progressDialog.setTitle("주문중입니다.");
-            progressDialog.show();
-
-        }
-
-
-        @Override
-        protected void onPostExecute(String result) {
-            super.onPostExecute(result);
-
-            Log.d(TAG, "POST response  - " + result);
-
-            if (result.contains("주문 목록에 추가했습니다.")) {
-
-                Toast.makeText(getApplicationContext(), "추가 완료되었습니다.", Toast.LENGTH_SHORT).show();
-
-            } else {
-
-                if(result.contains("Integrity constraint violation: 1062 Duplicate entry")) {
-
-                    Toast.makeText(getApplicationContext(), "이미 주문 목록에 존재하는 상품입니다!", Toast.LENGTH_SHORT).show();
-
-                } else {
-
-                    Toast.makeText(getApplicationContext(), "오류가 발생했습니다.", Toast.LENGTH_SHORT).show();
-
-                }
-            }
-
-            progressDialog.dismiss();
-
-
-        }
-
-
-        @Override
-        protected String doInBackground(String... params) {
-
-            String item_no = (String) params[1];
-            String email = (String) params[2];
-            String count = (String) params[3];
-            String class_name = (String) params[4];
-
-            String serverURL = (String) params[0];
-            String postParameters = "item_no=" + item_no + "&email=" + email + "&count=" + count + "&class_name=" + class_name;
-
-            try {
-
-                URL url = new URL(serverURL);
-                HttpURLConnection httpURLConnection = (HttpURLConnection) url.openConnection();
-
-
-                httpURLConnection.setReadTimeout(10000);
-                httpURLConnection.setConnectTimeout(10000);
-                httpURLConnection.setRequestMethod("POST");
-                httpURLConnection.connect();
-
-
-                OutputStream outputStream = httpURLConnection.getOutputStream();
-                outputStream.write(postParameters.getBytes("UTF-8"));
-                outputStream.flush();
-                outputStream.close();
-
-
-                int responseStatusCode = httpURLConnection.getResponseCode();
-                Log.d(TAG, "POST response code - " + responseStatusCode);
-
-                InputStream inputStream;
-                if (responseStatusCode == HttpURLConnection.HTTP_OK) {
-                    inputStream = httpURLConnection.getInputStream();
-                } else {
-                    inputStream = httpURLConnection.getErrorStream();
-                }
-
-
-                InputStreamReader inputStreamReader = new InputStreamReader(inputStream, "UTF-8");
-                BufferedReader bufferedReader = new BufferedReader(inputStreamReader);
-
-                StringBuilder sb = new StringBuilder();
-                String line = null;
-
-                while ((line = bufferedReader.readLine()) != null) {
-                    sb.append(line);
-                }
-
-
-                bufferedReader.close();
-
-
-                return sb.toString();
-
-
-            } catch (Exception e) {
-
-                Log.d(TAG, "InsertData: Error ", e);
-
-                return new String("Error: " + e.getMessage());
             }
 
         }
